Relocate Batcher to batcher/
