+++ differential-dataflow/tests/trace.rs
@@ -2,7 +2,8 @@ use timely::dataflow::operators::generic::OperatorInfo;
 use timely::progress::{Antichain, frontier::AntichainRef};
 
 use differential_dataflow::trace::implementations::{ValBatcher, ValSpine};
-use differential_dataflow::trace::{Description, Span, Trace, TraceReader, Batcher};
+use differential_dataflow::batcher::Batcher;
+use differential_dataflow::trace::{Description, Span, Trace, TraceReader};
 use differential_dataflow::trace::cursor::{Cursor, cursor_list};
 
 type IntegerTrace = ValSpine<u64, u64, usize, i64>;
